update: backend updated version

+++ src/main/java/com/example/demo/dto/CompositionDTO.java
@@ -9,5 +9,5 @@ public record CompositionDTO(
 
         @NotNull(message = "Quantity is required")
         @Positive(message = "Quantity must be positive")
-        Integer quantity
+        double quantity
 ) {}
