+++ src/main/java/com/example/demo/dto/CompositionResponseDTO.java
@@ -3,5 +3,5 @@ package com.example.demo.dto;
 public record CompositionResponseDTO(
         Long materialId,
         String materialName,
-        Integer quantity
+        double quantity
 ) {}
